refactor request types and storage
