+++ node/src/coordinator.rs
@@ -1,4 +1,4 @@
-use crate::ckd_request::CKDRequestStorage;
+use crate::storage::CKDRequestStorage;
 use crate::config::{ConfigFile, MpcConfig, ParticipantsConfig, SecretsConfig};
 use crate::db::{DBCol, SecretDB, EPOCH_ID_KEY};
 use crate::indexer::handler::ChainBlockUpdate;
@@ -21,7 +21,7 @@ use crate::providers::ckd::CKDProvider;
 use crate::providers::eddsa::{EddsaSignatureProvider, EddsaTaskId};
 use crate::providers::{EcdsaSignatureProvider, EcdsaTaskId};
 use crate::runtime::AsyncDroppableRuntime;
-use crate::sign_request::SignRequestStorage;
+use crate::storage::SignRequestStorage;
 use crate::tracking::{self};
 use crate::web::DebugRequest;
 use futures::future::BoxFuture;
