+++ node/src/indexer/fake.rs
@@ -6,7 +6,7 @@ use crate::config::ParticipantsConfig;
 use crate::indexer::handler::CKDRequestFromChain;
 use crate::indexer::types::ChainCKDRespondArgs;
 use crate::requests::recent_blocks_tracker::tests::TestBlockMaker;
-use crate::sign_request::SignatureId;
+use crate::types::SignatureId;
 use crate::tracking::{AutoAbortTask, AutoAbortTaskCollection};
 use anyhow::Context;
 use mpc_contract::config::Config;
