+++ node/src/providers/ecdsa/mod.rs
@@ -16,7 +16,8 @@ use crate::db::SecretDB;
 use crate::network::{MeshNetworkClient, NetworkTaskChannel};
 use crate::primitives::{MpcTaskId, UniqueId};
 use crate::providers::{PublicKeyConversion, SignatureProvider};
-use crate::sign_request::{SignRequestStorage, SignatureId};
+use crate::storage::SignRequestStorage;
+use crate::types::SignatureId;
 use crate::tracking;
 use anyhow::Context;
 use borsh::{BorshDeserialize, BorshSerialize};
