+++ node/src/providers/eddsa/mod.rs
@@ -7,7 +7,8 @@ use crate::config::{ConfigFile, MpcConfig, ParticipantsConfig};
 use crate::network::{MeshNetworkClient, NetworkTaskChannel};
 use crate::primitives::MpcTaskId;
 use crate::providers::{PublicKeyConversion, SignatureProvider};
-use crate::sign_request::{SignRequestStorage, SignatureId};
+use crate::storage::SignRequestStorage;
+use crate::types::SignatureId;
 use borsh::{BorshDeserialize, BorshSerialize};
 use mpc_contract::primitives::domain::DomainId;
 use mpc_contract::primitives::key_state::KeyEventId;
