+++ node/src/providers/mod.rs
@@ -13,7 +13,7 @@ pub mod eddsa;
 use crate::config::ParticipantsConfig;
 use crate::network::NetworkTaskChannel;
 use crate::primitives::{MpcTaskId, ParticipantId};
-use crate::sign_request::SignatureId;
+use crate::types::SignatureId;
 pub use ecdsa::EcdsaSignatureProvider;
 pub use ecdsa::EcdsaTaskId;
 use std::str::FromStr;
